requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.7.0
click==8.1.5
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.3
flake8==6.0.0
importlib-resources==6.0.0
inflection==0.5.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.8.1
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
rpds-py==0.8.10
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.16.2
